+++ CHANGELOG.md
@@ -4,18 +4,25 @@
     - Removed `ImageCacheProxy` API.
     - Added similar `ImagesExtension` API.
     - All *proxy* behavior can be easy converted to new API.
+    - SVG image extension now fully integrated.
 
 * **Breaking** Length units `max`, `min` and `abs` now take by value.
 * Fix memory leak in `Var::flat_map_vec` when the returned item var is a retained clone.
 
 * **Breaking** `FileDialogFilters::push_filter` now accepts any `IntoIterator<Item=str>`.
 
+* Add support for multi data clipboards in the view process API.
+    - **Breaking** View process API updated to take and return vectors of data.
+    - Not implemented in the surface API, this will happen in a future non breaking release.
+
 * Unify view-process capabilities info.
     - **Breaking** Removed `IMAGES.available_encoders/decoders`.
     - Added `IMAGES.available_formats` with `ImageFormat`.
     - **Breaking** View-process API event inited now has an `ViewProcessInfo` that contains implementation details.
     - **Breaking** Added `VIEW_PROCESS.info` and removed methods for specific capabilities.
-    - **Breaking** 
+    - Added `WINDOWS.available_operations` with what window operations the view-process implements.
+    - Added `DIALOG.available_native_dialogs` with what native dialogs the view-process implements.
+    - Added `CLIPBOARD.available_types` with what data types can be read/write by the view-process implementation.
 
 * Fix `ImageVar` updates not propagating in some complex bindings.